rename piechart file to match component
