+++ react-front-end/src/components/LeftDatas/LeftData.js
@@ -1,11 +1,12 @@
 import React from 'react';
 
+//TODO replace tempWords with useState of words that come in during stream
 import words from './tempWords';
 
 import './LeftData.scss';
 import Wordcloud from './Wordcloud';
 import SentiBarChart from './BarChart';
-import SentiPieChart from './PieChart';
+import SentiPieChart from './SentiPieChart';
 
 export default function LeftData() {
